+++ client/src/compnents/PreviewBox/index.tsx
@@ -1,15 +1,17 @@
+import IDB_File from "../../types/iDB_file";
 import Tile from "./Tile";
 import "./style.scss";
 
 export interface PreviewBoxProps {
-  files: File[] | null;
+  files: IDB_File[] | null;
+  removeCallback?: (fileID: string) => void;
 }
 
 export const PreviewBox = (props: PreviewBoxProps) => {
   return (
     <div className="preview-box">
       {props.files?.map((file, index) => (
-        <Tile key={index} file={file} />
+        <Tile key={index} file={file} removeCallback={props.removeCallback} />
       ))}
     </div>
   );